[libc][NFC] Integrate `FloatProperties` into `FPBits`
`FloatProperties` is always included when `FPBits` is. This will help further refactoring.
